mutation to change notification preferences
