+++ src/types/lernfair/NotificationPreferences.ts
@@ -1,0 +1,7 @@
+export type NotificationPreferences = {
+  [category: string]: PreferencesType
+}
+
+export type PreferencesType = {
+  [channel: string]: boolean
+}
